v1.0.0 - Code quality fixes and SubHardware support
- Fix SubHardware visibility in MainWindow.xaml (SuperIO sensors now display)
- Add logging in empty catch blocks (Logger, RelayCommand, WindowService)
- Add settings.json size validation (max 1MB)
- Change BackdropStyle from int to enum for type safety
- Use distinct icons for CPU/GPU/Motherboard
- Remove HardwareCategoryExtensions.GetIcon() duplication
- Decouple HardwareReference from HardwareNode model to HardwareService
- Update tests for BackdropStyle enum changes

+++ HardwareMonitorWinUI3.Tests/AppSettingsTests.cs
@@ -15,7 +15,7 @@ public class AppSettingsTests
         Assert.Equal(1200, settings.WindowWidth);
         Assert.Equal(800, settings.WindowHeight);
         Assert.Equal(250, settings.RefreshInterval);
-        Assert.Equal((int)BackdropStyle.MicaAlt, settings.BackdropStyle);
+        Assert.Equal(BackdropStyle.MicaAlt, settings.BackdropStyle);
         Assert.True(settings.ShowCPU);
         Assert.True(settings.ShowGPU);
         Assert.True(settings.ShowMotherboard);
